+++ storage-resize-images/jest.config.js
@@ -5,4 +5,5 @@ module.exports = {
   displayName: packageJson.name,
   rootDir: "./",
   preset: "ts-jest",
+  testMatch: ["**/__tests__/*.test.ts"],
 };
